Do not hardcode the name for verification test
The current test implementation expects "android" to be a part of the
webcam name. This holds true for the current devices that support
webcam but vendors have shown interest in changing the name.

This CL removes the name matching logic and instead uses udevadm to
get the serial of the device under test. This allows the vendors to
set any name they see fit.

Bug: 277159494
Flag: EXEMPT. Test fix.
Test: Locally tested to be working. Verified by vendors.
Change-Id: I01c566caf9cebcd32746d5a1defc0691ed26da22