Update test instructions
This CL updates the README.md with the most up to date instructions,
on how the webcam verification test should be run, and reflects the
fact that tests on Windows and MacOS are now deprecated.

README.md now also acts as the plaintext source for
DeviceAswebcaminstructions.pdf to prevent instructions from being
available in binary form only.

Bug: 277159494
Test: Instruction changes only.
Flag: EXEMPT. Test only change.
Change-Id: I35e98a9345f6613cd3e9881cc079d9d867a9426b