test: Remove dependence on adb root
The current implementation of the host script requires adb root to
test if the device supports uvc gadget mode or not. Unfortunately this
requirement makes the test unsuitable for user builds where root is
disabled.

With Ib124b755be55, CtsVerifier uses a test API to check for uvc
gadget support, and skip the test early. That allows the host scripts
to assume that the DUT will always support UVC.

This CL removes the logic to report UVC support thereby removing
reliance on adb root.

Bug: 434155777
Test: Verified that CtsVerifier can still fetch and skip Webcam test
      if uvc is not supported. No change for when uvc is supported.
Flag: EXEMPT. Test only change.
Change-Id: I4362bdd7022a61ca7f6bf19f619f8a2e541a393a