Use VIDIOC_(S/G)_PARMS to set frame interval in UVC gadget driver.
https://lore.kernel.org/all/20240403-uvc_request_length_by_interval-v7-4-e224bb1035f0@pengutronix.de/
introduced userspace breaking changes in the gadget driver. After this CL, if a userspace client
doesn't call the VIDIOC_S_PARAM ioctl with the frame interval set, it will always default
to 15 fps, which is undesirable and may not be in sync with what devices
advertise as supported frame rates.

Bug: 437154969
Bug: 438254326

Test: manual, frame rate isn't stuck at 15fps

Flag: EXEMPT; Bugfix

Change-Id: I4c770ac6e7260df31f03ff96f7b8628086bbc09d
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>