Export DeviceAsWebcamPreview activity.
For the webcam compliance test to be able to pull up the
DeviceAsWebcamPreview activity to verify correctness, DeviceAsWebcam's
AndroidManifest should export the Preview activity. Not doing so
results in a SecurityException when adb attempts to bring up the
preview activity.

This patch exports the DeviceAsWebcamPreview activity in
AndroidManifest.xml to ensure the compliance test does not need to
rely on adb's root access.

Bug: 437352393
Test: Verified that adb can pull up the activity without root
      privileges.
Flag: EXEMPT. Bug fix.
Change-Id: I90bc2caad49271bd117387711699ee190cfadc8b