Force webcam FPS test to pass on older vendor API levels.
This change introduces a `FORCE_PASS` result for the webcam test. If the
FPS test fails, it will still report be reported as `FORCE_PASS` when
the device's `ro.vendor.api_level` is 202504 or earlier. This is done to
accommodate devices affected by a past test bug, where the test would
pass unconditionally.

Bug: 441160351
Test: Manually verified that devices launching with first API level
202504 is force passed even on failures
Flag: EXEMPT TEST_ONLY
Change-Id: Ia89471c2cbf423307ab738c4467cbdb3b532868f