Set SurfaceTexture buffer size when starting preview stream.
When starting the preview stream, it is possible that the passed
SurfaceTexture is newly created and therefore does not have the correct
size set. This can occasionally lead to the preview being stretched when
the DeviceAsWebcamPreview Activity is newly created.

This CL updates CameraController's startPreviewStreaming method to
update the SurfaceTexture's buffer size to the current preview size
before starting the configuring the preview stream.

Bug: 447206947
Test: Manually verified that the stretching is no longer observable across 15 runs.
Flag: EXEMPT BUGFIX
Change-Id: Iec88c23fafe4d9d749492f5a53311a7fa82ef8bf